+++ extension/server/src/providers/hover.ts
@@ -1,8 +1,8 @@
 import { Hover, HoverParams, MarkupKind, Range } from 'vscode-languageserver';
 import { documents, getReturnValue, getWordRangeAtPosition, parser, prettyKeywords } from '.';
-import Parser from "../../../../language/parser";
+import Parser from '../../../../language/ile/parser';
 import { URI } from 'vscode-uri';
-import { Keywords } from '../../../../language/parserTypes';
+import { Keywords } from '../../../../language/ile/parserTypes';
 import Declaration from '../../../../language/models/declaration';
 
 export default async function hoverProvider(params: HoverParams): Promise<Hover | undefined> {
@@ -125,17 +125,15 @@ export default async function hoverProvider(params: HoverParams): Promise<Hover 
 						const foundUri = URI.parse(include.uri);
 
 						if (foundUri.scheme === `member`) {
-							const lastIndex = foundUri.path.lastIndexOf(`.`);
+							const lastIndex = foundUri.fsPath.lastIndexOf(`.`);
 							if (lastIndex >= 0) {
-								displayName = foundUri.path.substring(0, lastIndex);
+								displayName = foundUri.fsPath.substring(0, lastIndex);
 							} else {
-								displayName = foundUri.path;
+								displayName = foundUri.fsPath;
 							}
 
-							if (displayName.startsWith(`/`)) displayName = displayName.substring(1);
-
 						} else {
-							displayName = foundUri.path;
+							displayName = foundUri.fsPath;
 						}
 					}